+++ source/Modules/SR_ShardManager/src/ShardManager.h
@@ -11,7 +11,7 @@ namespace ShardManager
     class CShardManager {
     public:
         static void Initialize();
-        BYTE GetMyServerBodyID();
+        WORD GetMyServerBodyID();
         HWND GetWinHandle();
         void SendMsgToServerBody(E_MODULE::Type Type, CMsg* pMsg);
